Add synopsis check
The check whether the SYNOPSIS is auto-generated only works if there is no help. If there is help and the SYNOPSIS tag is left out your test comes back as passed.

I added a second check to make sure that if there is help, that synopsis is not empty.

The two checks should catch all possible unwanted states Synopsis can be in.